+++ CHANGES.md
@@ -1,5 +1,6 @@
 # 0.2
 
+- add `Primes.prime_seq`
 - lower the minimal supported version of OCaml to 4.07
 
 # 0.1